# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ULTRA GURU Fork Verifier API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ForkCheckInput {
  /**
     * GitHub username to check
     * @minLength 1
     */
  username: string;
}

export interface ForkCheckResult {
  /** Whether the user has forked the repo */
  hasFork: boolean;
  username: string;
  /**
     * URL of the user's fork if it exists
     * @nullable
     */
  forkUrl?: string | null;
  /**
     * Heroku deploy URL to redirect to if verified
     * @nullable
     */
  redirectUrl?: string | null;
}

export interface ForkCountResult {
  /** Number of forks on GuruhTech/ULTRA-GURU */
  count: number;
}

export interface ErrorResponse {
  error: string;
}