# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ULTRA GURU Fork Verifier API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: verify
    description: Fork verification
  - name: stats
    description: Repository statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /fork-count:
    get:
      operationId: getForkCount
      tags: [stats]
      summary: Get fork count
      description: Returns the current number of forks for the ULTRA-GURU repo
      responses:
        "200":
          description: Fork count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForkCountResult"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verify-fork:
    post:
      operationId: verifyFork
      tags: [verify]
      summary: Verify GitHub fork
      description: Check if a GitHub user has forked the ULTRA-GURU repo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForkCheckInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForkCheckResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ForkCheckInput:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          description: GitHub username to check
          minLength: 1
    ForkCheckResult:
      type: object
      required:
        - hasFork
        - username
      properties:
        hasFork:
          type: boolean
          description: Whether the user has forked the repo
        username:
          type: string
        forkUrl:
          type: ["string", "null"]
          description: URL of the user's fork if it exists
        redirectUrl:
          type: ["string", "null"]
          description: Heroku deploy URL to redirect to if verified
    ForkCountResult:
      type: object
      required:
        - count
      properties:
        count:
          type: integer
          description: Number of forks on GuruhTech/ULTRA-GURU
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string