# lib/api-zod/src/generated/types/forkCountResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ULTRA GURU Fork Verifier API
 * OpenAPI spec version: 0.1.0
 */

export interface ForkCountResult {
  /** Number of forks on GuruhTech/ULTRA-GURU */
  count: number;
}